Add an onRejected promise handler for to getChildren
We came across a scenario when using the px-context-browser where the call to getChildren would internally get a server error due to a timeout retrieving assets from APM.  This appears to be a regular occurrence.
Our immediate problem was that the spinner would never go away until we refreshed the page.
We have temporarily handled this ourselves by having our getChildren return an empty array of children, while keeping the existing parent context.  However this pushed our error condition through the success functionality in getNewChildren.  A better solution would be to have the getNewChildren be able to handle the error condition, and log a warning as it appears that many other clients would come across this scenario.

We have tested this with our implementation, and it allows us to simply allow the user to retry getting the children, while knowing why the error happened in the first place.