fix bug where data-binding 'opened' into context browser breaks active reset
This bug was found while building the demo page for px-context-browser. When a selectedRoute was set through data-binding and opened was set to false through data-binding, the context browser would correctly open to that item, but then when the user navigated away from that item, closed the context browser and quickly reopened it in less than 5 sec, they would be brought back to the selected item rather than the view they were just looking at.